backend/pyproject.toml
[project]
name = "marketseer-backend"
version = "0.1.0"
description = "MarketSeer 金融数据采集与分析系统后端"
authors = [{ name = "MarketSeer Team" }]
requires-python = ">=3.11,<3.14"
dependencies = [
    # web framework
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=4.0.0",
    "flask-migrate>=4.0.0",
    "sqlalchemy>=2.0.0",

    # PostgreSQL driver
    "psycopg2-binary>=2.9.9",

    # financial data libraries
    "akshare>=1.17.79",
    "efinance>=0.5.0",
    "yfinance>=0.2.0",

    # data manipulation
    "pandas>=2.0.0",
    "numpy>=2.3.4",
    "python-dotenv>=1.0.0",
    "exchange-calendars>=4.0.0",

    # scheduler
    "apscheduler>=3.10.0",

    # logger
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",

    # Jupyter support
    "notebook>=7.0.0",
    "ipykernel>=7.0.0",
    "ipywidgets>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = ["ipykernel>=7.1.0"]
